Config build and import modules
