+++ .eslintignore
@@ -1,0 +1,2 @@
+webpack.config.js
+public/
