+++ .gitignore
@@ -1,12 +1,10 @@
-* text eol=lf
+.DS_Store
+Thumbs.db
 
-*.png binary
-*.jpg binary
-*.jpeg binary
-*.webp binary
-*.woff binary
-*.woff2 binary
-
-node_modules
 .vscode
 
+*.log*
+node_modules/
+
+public/bundle.js
+public/bundle.js.map
