+++ src/main.js
@@ -1,288 +1,14 @@
-"use strict";
+import {createSiteMenuTemplate} from "./view/site-menu.js";
+import {createFilterTemplate} from "./view/filter.js";
+import {createSortingTemplate} from "./view/sorting.js";
+import {createBoardTemplate} from "./view/board.js";
+import {createTaskTemplate} from "./view/task.js";
+import {createLoadMoreButtonTemplate} from "./view/load-more-button.js";
+import {createEditTaskTemplate} from "./view/edit-task.js";
 
 const TASK_AMOUNT = 3;
 
-const createSiteMenuTemplate = () => {
-  return (
-    `<section class="control__btn-wrap">
-      <input
-        type="radio"
-        name="control"
-        id="control__new-task"
-        class="control__input visually-hidden"
-      />
-      <label for="control__new-task" class="control__label control__label--new-task"
-        >+ ADD NEW TASK</label
-      >
-      <input
-        type="radio"
-        name="control"
-        id="control__task"
-        class="control__input visually-hidden"
-        checked
-      />
-      <label for="control__task" class="control__label">TASKS</label>
-      <input
-        type="radio"
-        name="control"
-        id="control__statistic"
-        class="control__input visually-hidden"
-      />
-      <label for="control__statistic" class="control__label"
-        >STATISTICS</label
-      >
-    </section>`
-  );
-};
-
-const createFilterTemplate = () => {
-  return (
-    `<section class="main__filter filter container">
-      <input
-        type="radio"
-        id="filter__all"
-        class="filter__input visually-hidden"
-        name="filter"
-        checked
-      />
-      <label for="filter__all" class="filter__label">
-        All <span class="filter__all-count">13</span></label
-      >
-      <input
-        type="radio"
-        id="filter__overdue"
-        class="filter__input visually-hidden"
-        name="filter"
-        disabled
-      />
-      <label for="filter__overdue" class="filter__label"
-        >Overdue <span class="filter__overdue-count">0</span></label
-      >
-      <input
-        type="radio"
-        id="filter__today"
-        class="filter__input visually-hidden"
-        name="filter"
-        disabled
-      />
-      <label for="filter__today" class="filter__label"
-        >Today <span class="filter__today-count">0</span></label
-      >
-      <input
-        type="radio"
-        id="filter__favorites"
-        class="filter__input visually-hidden"
-        name="filter"
-      />
-      <label for="filter__favorites" class="filter__label"
-        >Favorites <span class="filter__favorites-count">1</span></label
-      >
-      <input
-        type="radio"
-        id="filter__repeating"
-        class="filter__input visually-hidden"
-        name="filter"
-      />
-      <label for="filter__repeating" class="filter__label"
-        >Repeating <span class="filter__repeating-count">1</span></label
-      >
-      <input
-        type="radio"
-        id="filter__archive"
-        class="filter__input visually-hidden"
-        name="filter"
-      />
-      <label for="filter__archive" class="filter__label"
-        >Archive <span class="filter__archive-count">115</span></label
-      >
-    </section>`
-  );
-};
-
-const createSortingTemplate = () => {
-  return (
-    `<div class="board__filter-list">
-      <a href="#" class="board__filter">SORT BY DEFAULT</a>
-      <a href="#" class="board__filter">SORT BY DATE up</a>
-      <a href="#" class="board__filter">SORT BY DATE down</a>
-    </div>`
-  );
-};
-
-const createBoardTemplate = () => {
-  return (
-    `<section class="board container">
-      <div class="board__tasks"></div>
-    </section>`
-  );
-};
-
-const createTaskTemplate = () => {
-  return (
-    `<article class="card card--black">
-      <div class="card__form">
-        <div class="card__inner">
-          <div class="card__control">
-            <button type="button" class="card__btn card__btn--edit">
-              edit
-            </button>
-            <button type="button" class="card__btn card__btn--archive">
-              archive
-            </button>
-            <button type="button" class="card__btn card__btn--favorites card__btn--disabled">
-              favorites
-            </button>
-          </div>
-
-          <div class="card__color-bar">
-            <svg class="card__color-bar-wave" width="100%" height="10">
-              <use xlink:href="#wave"></use>
-            </svg>
-          </div>
-
-          <div class="card__textarea-wrap">
-            <p class="card__text">Example default task with default color.</p>
-          </div>
-
-          <div class="card__settings">
-            <div class="card__details">
-              <div class="card__dates">
-                <div class="card__date-deadline">
-                  <p class="card__input-deadline-wrap">
-                    <span class="card__date">23 September</span>
-                  </p>
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-    </article>`
-  );
-};
-
-const createRepeatDayTemplate = (day, active) => {
-  return (
-    `<input
-      class="visually-hidden card__repeat-day-input"
-      type="checkbox"
-      name="repeat"
-      value="${day}"
-      id="repeat-${day}"
-      ${active ? `checked` : ``}
-    />
-    <label class="card__repeat-day" for="repeat-${day}"
-      >${day}</label
-    >`
-  );
-};
-
-const createColorTaskTemplate = (color, active) => {
-  return (
-    `<input
-      type="radio"
-      id="color-${color}"
-      class="card__color-input card__color-input--${color} visually-hidden"
-      name="color"
-      value="${color}"
-      ${active ? `checked` : ``}
-    />
-    <label
-      for="color-${color}"
-      class="card__color card__color--${color}"
-      >${color}</label
-    >`
-  );
-};
-
-const createEditTaskTemplate = () => {
-  return (
-    `<article class="card card--edit card--yellow card--repeat">
-      <form class="card__form" method="get">
-        <div class="card__inner">
-          <div class="card__color-bar">
-            <svg class="card__color-bar-wave" width="100%" height="10">
-              <use xlink:href="#wave"></use>
-            </svg>
-          </div>
-
-          <div class="card__textarea-wrap">
-            <label>
-              <textarea
-                class="card__text"
-                placeholder="Start typing your text here..."
-                name="text"
-              >This is example of task edit. You can set date and chose repeating days and color.</textarea>
-            </label>
-          </div>
-
-          <div class="card__settings">
-            <div class="card__details">
-              <div class="card__dates">
-                <button class="card__date-deadline-toggle" type="button">
-                  date: <span class="card__date-status">yes</span>
-                </button>
-
-                <fieldset class="card__date-deadline">
-                  <label class="card__input-deadline-wrap">
-                    <input
-                      class="card__date"
-                      type="text"
-                      placeholder=""
-                      name="date"
-                      value="23 September 16:15"
-                    />
-                  </label>
-                </fieldset>
-
-                <button class="card__repeat-toggle" type="button">
-                  repeat:<span class="card__repeat-status">yes</span>
-                </button>
-
-                <fieldset class="card__repeat-days">
-                  <div class="card__repeat-days-inner">
-                    ${createRepeatDayTemplate(`mo`, false)}
-                    ${createRepeatDayTemplate(`tu`, true)}
-                    ${createRepeatDayTemplate(`we`, false)}
-                    ${createRepeatDayTemplate(`th`, false)}
-                    ${createRepeatDayTemplate(`fr`, true)}
-                    ${createRepeatDayTemplate(`sa`, false)}
-                    ${createRepeatDayTemplate(`su`, true)}
-                  </div>
-                </fieldset>
-              </div>
-            </div>
-
-            <div class="card__colors-inner">
-              <h3 class="card__colors-title">Color</h3>
-              <div class="card__colors-wrap">
-                ${createColorTaskTemplate(`black`, false)}
-                ${createColorTaskTemplate(`yellow`, true)}
-                ${createColorTaskTemplate(`blue`, false)}
-                ${createColorTaskTemplate(`green`, false)}
-                ${createColorTaskTemplate(`pink`, false)}
-              </div>
-            </div>
-          </div>
-
-          <div class="card__status-btns">
-            <button class="card__save" type="submit">save</button>
-            <button class="card__delete" type="button">delete</button>
-          </div>
-        </div>
-      </form>
-    </article>`
-  );
-};
-
-const createLoadMoreButtonTemplate = () => {
-  return (
-    `<button class="load-more" type="button">load more</button>`
-  );
-};
-
-
-const render = (container, template, place) => {
+export const render = (container, template, place) => {
   container.insertAdjacentHTML(place, template);
 };
